Move oci to daemon/pkg/oci
