+++ daemon/builder/dockerfile/evaluator.go
@@ -27,8 +27,8 @@ import (
 
 	"github.com/docker/docker/daemon/builder"
 	"github.com/docker/docker/daemon/internal/image"
+	"github.com/docker/docker/daemon/pkg/oci"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/oci"
 	"github.com/moby/buildkit/frontend/dockerfile/instructions"
 	"github.com/moby/buildkit/frontend/dockerfile/shell"
 	"github.com/moby/moby/api/types/container"
