+++ daemon/container.go
@@ -19,10 +19,10 @@ import (
 	"github.com/docker/docker/daemon/container"
 	"github.com/docker/docker/daemon/internal/image"
 	"github.com/docker/docker/daemon/network"
+	"github.com/docker/docker/daemon/pkg/oci/caps"
 	"github.com/docker/docker/daemon/pkg/opts"
 	volumemounts "github.com/docker/docker/daemon/volume/mounts"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/oci/caps"
 	"github.com/docker/go-connections/nat"
 	containertypes "github.com/moby/moby/api/types/container"
 	"github.com/moby/moby/api/types/mount"
