+++ daemon/container/container.go
@@ -26,10 +26,10 @@ import (
 	"github.com/docker/docker/daemon/logger/local"
 	"github.com/docker/docker/daemon/logger/loggerutils/cache"
 	"github.com/docker/docker/daemon/network"
+	"github.com/docker/docker/daemon/pkg/oci"
 	"github.com/docker/docker/daemon/volume"
 	volumemounts "github.com/docker/docker/daemon/volume/mounts"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/oci"
 	"github.com/docker/go-units"
 	containertypes "github.com/moby/moby/api/types/container"
 	"github.com/moby/moby/api/types/events"
