+++ daemon/oci_windows.go
@@ -15,8 +15,8 @@ import (
 	"github.com/docker/docker/daemon/config"
 	"github.com/docker/docker/daemon/container"
 	"github.com/docker/docker/daemon/internal/image"
+	"github.com/docker/docker/daemon/pkg/oci"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/oci"
 	"github.com/moby/moby/api/types/backend"
 	containertypes "github.com/moby/moby/api/types/container"
 	"github.com/opencontainers/runtime-spec/specs-go"
