+++ daemon/pkg/plugin/v2/plugin_linux.go
@@ -9,9 +9,9 @@ import (
 	"runtime"
 	"strings"
 
+	"github.com/docker/docker/daemon/pkg/oci"
 	"github.com/docker/docker/internal/rootless/mountopts"
 	"github.com/docker/docker/internal/sliceutil"
-	"github.com/docker/docker/oci"
 	"github.com/moby/moby/api/types"
 	"github.com/moby/sys/userns"
 	"github.com/opencontainers/runtime-spec/specs-go"
