+++ integration/container/create_test.go
@@ -11,9 +11,9 @@ import (
 
 	containerd "github.com/containerd/containerd/v2/client"
 	cerrdefs "github.com/containerd/errdefs"
+	"github.com/docker/docker/daemon/pkg/oci"
 	testContainer "github.com/docker/docker/integration/internal/container"
 	net "github.com/docker/docker/integration/internal/network"
-	"github.com/docker/docker/oci"
 	"github.com/docker/docker/pkg/stringid"
 	"github.com/docker/docker/testutil"
 	"github.com/moby/moby/api/types/container"
